move getTOMLFieldValue to testutil
